CORE-39098 addressed pr comments
